Remove `forcePublish` flag for canary release
`lerna publish --canary` creates a version based on the changes that it finds from git. If there are no changes with `--forcePublish` it will try to publish a version that already exists, causing the command to fail.

While I don't agree with this behaviour of `lerna`, I also think we don't need to publish new canary versions, when there are no changes in the packages.