chore: replace cache action with setup-node cache (#3688)
Replace the manual caching config with the default config from setup-node. The major difference is, that the node_modules are not cached directly as a directory. Instead the dependencies are cached globally and then transferred to the node_modules on install. That way the turbo cache is ignored as well.