chore: Fix memory tests after v3 release (#3754)
This PR fixes the failing memory tests due to changes caused by the v3 release of the SDK.

The new baseline comparison is v2 instead of v1.

v2 is the baseline because it is not updated anymore, so it is a stable reference point to judge how memory usage changes in v3/canary.

Note that the old 90% threshold does not seem to work anymore, updated to 105%. This should at least let us know if we are getting worse in memory usage. Threshold may be updated if we improve memory usage in the future.